Moved all Magento files to mage/ subdirectories
