+++ src/eu/shooktea/vmsm/module/Magento.java
@@ -1,27 +1,23 @@
 package eu.shooktea.vmsm.module;
 
 import com.teamdev.jxbrowser.chromium.Browser;
-import com.teamdev.jxbrowser.chromium.LoadURLParams;
 import com.teamdev.jxbrowser.chromium.dom.*;
 import com.teamdev.jxbrowser.chromium.events.FinishLoadingEvent;
 import com.teamdev.jxbrowser.chromium.events.LoadAdapter;
 import com.teamdev.jxbrowser.chromium.events.LoadListener;
 import eu.shooktea.vmsm.Start;
 import eu.shooktea.vmsm.VirtualMachine;
-import eu.shooktea.vmsm.view.controller.MagentoConfig;
-import eu.shooktea.vmsm.view.controller.MagentoNewModule;
-import eu.shooktea.vmsm.view.controller.MagentoReportsList;
+import eu.shooktea.vmsm.view.controller.mage.MagentoConfig;
+import eu.shooktea.vmsm.view.controller.mage.MagentoNewModule;
+import eu.shooktea.vmsm.view.controller.mage.MagentoReportsList;
 import eu.shooktea.vmsm.view.controller.MainWindow;
 import javafx.application.Platform;
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
 import javafx.geometry.Orientation;
 import javafx.scene.Node;
 import javafx.scene.control.*;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.KeyCombination;
-import javafx.scene.web.WebEngine;
 import org.reactfx.value.Val;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
